requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
python-dotenv==1.0.1
google-generativeai==0.8.4
notion-client==2.2.1
python-pptx==1.0.2
pydantic==2.10.6
pydantic-settings==2.7.1
python-multipart==0.0.20
